Rename notification to solid-notifications
