+++ packages/core/package.json
@@ -13,7 +13,7 @@
   "dependencies": {
     "@activitypods/announcer": "1.1.0",
     "@activitypods/ontologies": "1.1.0",
-    "@activitypods/notification": "1.1.0",
+    "@activitypods/solid-notifications": "1.1.0",
     "@semapps/activitypub": "0.6.0-alpha.8",
     "@semapps/auth": "0.6.0-alpha.8",
     "@semapps/jsonld": "0.6.0-alpha.8",
