+++ packages/core/service.js
@@ -14,7 +14,7 @@ const { TripleStoreService } = require('@semapps/triplestore');
 const { WebAclService } = require('@semapps/webacl');
 const { WebfingerService } = require('@semapps/webfinger');
 const { WebIdService } = require('@semapps/webid');
-const { NotificationProviderService } = require('@activitypods/notification');
+const { NotificationProviderService } = require('@activitypods/solid-notifications');
 const { apods, interop, notify, oidc } = require('@activitypods/ontologies');
 const ApiService = require('./services/api');
 const AppOpenerService = require('./services/app-opener');
