+++ packages/solid-notifications/LICENSE
