+++ packages/solid-notifications/index.js
