+++ packages/solid-notifications/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@activitypods/notification",
+  "name": "@activitypods/solid-notifications",
   "version": "1.1.0",
   "description": "ActivityPods package for Solid notifications",
   "license": "Apache-2.0",
