+++ tests/initialize.js
@@ -10,7 +10,7 @@ const { apods, interop, oidc } = require('@semapps/ontologies');
 const { WebAclMiddleware } = require('@semapps/webacl');
 const { ObjectsWatcherMiddleware } = require('@semapps/sync');
 const { CoreService, interopContext } = require('@activitypods/core');
-const { NotificationListenerService } = require('@activitypods/notification');
+const { NotificationListenerService } = require('@activitypods/solid-notifications');
 const { CoreService: SemAppsCoreService } = require('@semapps/core');
 const { AnnouncerService } = require('@activitypods/announcer');
 const CONFIG = require('./config');
